+++ RELEASE-NOTES.md
@@ -9,6 +9,22 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 
 ## [Unreleased] - Release date yyyy-mm-dd
 
+## [Version 0.5.2] - Release date 2022-10-05
+
+### Added
+- Added `blt_convert_to_system_includes` macro to convert existing interface includes to system interface includes.
+- `blt_check_code_compiles` which compiles a C++ code snippet and returns the result.
+- Added variable ``BLT_CMAKE_IMPLICIT_LINK_LIBRARIES_EXCLUDE`` for filtering
+  link libraries implicitly added by CMake. See the following example host-config:
+  ``host-configs/llnl/blueos_3_ppc64le_ib_p9/clang@upstream_nvcc_xlf.cmake``
+
+### Changed
+- Added three extra options to `blt_print_target_properties` macro to print properties of
+  target's children as well as limit the properties printed with regular expressions:
+    - CHILDREN (true/ false) whether or not you want to print the target's children's properties as well (recursively)
+    - PROPERTY_NAME_REGEX (regular expression string) reduce which properties to print by name
+    - PROPERTY_VALUE_REGEX (regular expression string) reduce which properties to print by value
+
 ## [Version 0.5.1] - Release date 2022-04-22
 
 ### Added
@@ -21,6 +37,7 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
   CMAKE_CUDA_STANDARD if set otherwise falls back on BLT_CXX_STD or CMAKE_CXX_STANDARD.
 - Removed extra HIP offload flags that were being added as generator expressions as opposed to simple
   flags.
+- Skip check for valid `ELEMENTS` parameter in `blt_list_append` macro when not appending
 
 ### Removed
 - Removed support for deprecated HCC.
@@ -251,7 +268,8 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 
 
 
-[Unreleased]:    https://github.com/LLNL/blt/compare/v0.5.1...develop
+[Unreleased]:    https://github.com/LLNL/blt/compare/v0.5.2...develop
+[Version 0.5.2]: https://github.com/LLNL/blt/compare/v0.5.1...v0.5.2
 [Version 0.5.1]: https://github.com/LLNL/blt/compare/v0.5.0...v0.5.1
 [Version 0.5.0]: https://github.com/LLNL/blt/compare/v0.4.1...v0.5.0
 [Version 0.4.1]: https://github.com/LLNL/blt/compare/v0.4.0...v0.4.1